backend/requirements.txt
# ============================================
# Core Dependencies (기존 파이프라인)
# ============================================

# PDF 처리
PyPDF2==3.0.1
pdfplumber==0.10.3
PyMuPDF==1.23.8

# 벡터 DB (임베딩은 외부 BGE Embedding 서비스 사용)
qdrant-client==1.7.0

# LLM API
openai==1.6.1

# 유틸리티
python-dotenv==1.0.0
tqdm==4.66.1

# ============================================
# Backend API Dependencies (FastAPI)
# ============================================

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Async File Operations
aiofiles==23.2.1

# Session Management & Security (선택적)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP Client
httpx==0.26.0

# ============================================
# Development Dependencies (선택적)
# ============================================

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Code Quality
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.3
